My final submission for Project 2
This is my final code for Project 2. I'm so sorry that I haven't been updating my code. I was a bit struggle to work with different mode files on VSC, thus, I was so nervous to try to refine the code that I forgot about open up the phenakistoscope. I did create a different file to submit my code on GitHub, which I did not notice that it located in a different file that not connected to the classroom. The theme of my GIF is about the Asian mid-autumn festival, celebrated in many places in East Asia and in many Asian communities around the world. The festival typically coincides with a full moon around the time of the autumnal equinox. It is also called the Moon Festival or the Mooncake Festival.